+++ scripts/macros/2024/classFeatures/cleric/channelDivinity.js
@@ -103,9 +103,22 @@ async function damage({trigger, workflow}) {
     await workflow.setDamageRolls(workflow.damageRolls);
     workflow.defaultDamageType = selection;
 }
+async function added({trigger: {entity: item, actor}}) {
+    let classIdentifier = itemUtils.getConfig(item, 'classIdentifier');
+    let scaleIdentifier = itemUtils.getConfig(item, 'scaleIdentifier');
+    let scale = actor.system.scale?.[classIdentifier]?.[scaleIdentifier];
+    if (scale) return;
+    let classItem = actor.classes[classIdentifier];
+    if (!classItem) return;
+    let classData = genericUtils.duplicate(classItem.toObject());
+    classData.system.advancement.push(channelDivinity.scales[0].data);
+    await genericUtils.update(classItem, {'system.advancement': classData.system.advancement});
+    let message = genericUtils.format('CHRISPREMADES.Requirements.ScaleAdded', {classIdentifier, scaleIdentifier});
+    genericUtils.notify(message, 'info');
+}
 export let channelDivinity = {
     name: 'Channel Divinity',
-    version: '1.2.12',
+    version: '1.2.13',
     rules: 'modern',
     midi: {
         item: [
@@ -135,6 +148,12 @@ export let channelDivinity = {
             }
         ]
     },
+    createItem: [
+        {
+            pass: 'created',
+            macro: added
+        }
+    ],
     config: [
         {
             value: 'creatureTypes',