+++ CHANGELOG.md
@@ -1,3 +1,14 @@
+# 1.2.13 Release:
+## New Modern Content:
+- Disciple of Life
+- Preserve Life
+## Bug Fixes:
+- Item Medkit
+- Channel Divinity
+## API Updates:
+- The combat event will now pass along the turn information to the trigger.
+- Templates will now be flagged with the rules version of the item they're tied to, when relevent.
+
 # 1.2.12 Release:
 ## New Modern Content:
 - Divine Order: Thaumaturge